streaming modes core functionality implemented
implemented. Has a few problems, when streaming mode on, and play a song, and open popup.html, it goes to normal speed for 1 second then goes back to streaming mode. When use streaming mode and change songs, in popup it doesnt show in the bar the selected playback (it is blank field)